REFACTOR `random()` test method to `RandomTest`
